refactor: extract step-execution mixin from engine/plan_execute_loop.py
