+++ src/synthorg/engine/plan_execute_step_mixin.py
@@ -1,0 +1,214 @@
+"""Step-execution mixin for :class:`PlanExecuteLoop`.
+
+Extracts step-turn running, completion handling, tool-call dispatch,
+checkpoint invocation, and result finalization into a mixin so the
+top-level loop file stays under the project's size limit.
+"""
+
+import copy
+from typing import TYPE_CHECKING
+
+from synthorg.engine.loop_helpers import (
+    call_provider,
+    check_budget,
+    check_response_errors,
+    check_shutdown,
+    classify_turn,
+    clear_last_turn_tool_calls,
+    execute_tool_calls,
+    make_turn_record,
+    response_to_message,
+)
+from synthorg.engine.loop_protocol import (
+    BudgetChecker,
+    ExecutionResult,
+    ShutdownChecker,
+    TurnRecord,
+)
+from synthorg.engine.plan_helpers import assess_step_success
+from synthorg.engine.plan_models import ExecutionPlan  # noqa: TC001
+from synthorg.observability import get_logger
+from synthorg.observability.events.execution import (
+    EXECUTION_CHECKPOINT_CALLBACK_FAILED,
+    EXECUTION_LOOP_TURN_COMPLETE,
+    EXECUTION_PLAN_STEP_TRUNCATED,
+)
+from synthorg.providers.enums import FinishReason
+
+if TYPE_CHECKING:
+    from synthorg.engine.approval_gate import ApprovalGate
+    from synthorg.engine.checkpoint.callback import CheckpointCallback
+    from synthorg.engine.context import AgentContext
+    from synthorg.providers.models import (
+        CompletionConfig,
+        CompletionResponse,
+        ToolDefinition,
+    )
+    from synthorg.providers.protocol import CompletionProvider
+    from synthorg.tools.invoker import ToolInvoker
+
+logger = get_logger(__name__)
+
+
+class PlanExecuteStepMixin:
+    """Mixin providing the step-execution helpers for PlanExecuteLoop."""
+
+    _approval_gate: ApprovalGate | None
+    _checkpoint_callback: CheckpointCallback | None
+
+    async def _run_step_turn(  # noqa: PLR0913
+        self,
+        ctx: AgentContext,
+        provider: CompletionProvider,
+        model: str,
+        config: CompletionConfig,
+        tool_defs: list[ToolDefinition] | None,
+        tool_invoker: ToolInvoker | None,
+        turns: list[TurnRecord],
+        budget_checker: BudgetChecker | None,
+        shutdown_checker: ShutdownChecker | None,
+    ) -> AgentContext | ExecutionResult | tuple[AgentContext, bool]:
+        """Execute a single turn within a step's mini-ReAct sub-loop."""
+        shutdown_result = check_shutdown(ctx, shutdown_checker, turns)
+        if shutdown_result is not None:
+            return shutdown_result
+        budget_result = check_budget(ctx, budget_checker, turns)
+        if budget_result is not None:
+            return budget_result
+
+        turn_number = ctx.turn_count + 1
+        response = await call_provider(
+            ctx,
+            provider,
+            model,
+            tool_defs,
+            config,
+            turn_number,
+            turns,
+        )
+        if isinstance(response, ExecutionResult):
+            return response
+
+        turns.append(
+            make_turn_record(
+                turn_number,
+                response,
+                call_category=classify_turn(turn_number, response, ctx),
+                provider_metadata=response.provider_metadata,
+            )
+        )
+
+        error = check_response_errors(ctx, response, turn_number, turns)
+        if error is not None:
+            return error
+
+        ctx = ctx.with_turn_completed(
+            response.usage,
+            response_to_message(response),
+        )
+        logger.info(
+            EXECUTION_LOOP_TURN_COMPLETE,
+            execution_id=ctx.execution_id,
+            turn=turn_number,
+            finish_reason=response.finish_reason.value,
+            tool_call_count=len(response.tool_calls),
+        )
+
+        await self._invoke_checkpoint_callback(ctx, turn_number)
+
+        if not response.tool_calls:
+            return self._handle_step_completion(ctx, response, turn_number)
+
+        return await self._handle_step_tool_calls(
+            ctx,
+            tool_invoker,
+            response,
+            turn_number,
+            turns,
+            shutdown_checker,
+        )
+
+    def _handle_step_completion(
+        self,
+        ctx: AgentContext,
+        response: CompletionResponse,
+        turn_number: int,
+    ) -> tuple[AgentContext, bool]:
+        """Assess step success and log truncation if applicable."""
+        success = assess_step_success(response)
+        if response.finish_reason == FinishReason.MAX_TOKENS:
+            logger.warning(
+                EXECUTION_PLAN_STEP_TRUNCATED,
+                execution_id=ctx.execution_id,
+                turn=turn_number,
+                truncated=True,
+            )
+        return ctx, success
+
+    async def _handle_step_tool_calls(  # noqa: PLR0913
+        self,
+        ctx: AgentContext,
+        tool_invoker: ToolInvoker | None,
+        response: CompletionResponse,
+        turn_number: int,
+        turns: list[TurnRecord],
+        shutdown_checker: ShutdownChecker | None,
+    ) -> AgentContext | ExecutionResult:
+        """Check shutdown and execute tool calls for a step turn."""
+        shutdown_result = check_shutdown(ctx, shutdown_checker, turns)
+        if shutdown_result is not None:
+            clear_last_turn_tool_calls(turns)
+            return shutdown_result.model_copy(
+                update={"turns": tuple(turns)},
+            )
+
+        return await execute_tool_calls(
+            ctx,
+            tool_invoker,
+            response,
+            turn_number,
+            turns,
+            approval_gate=self._approval_gate,
+        )
+
+    async def _invoke_checkpoint_callback(
+        self,
+        ctx: AgentContext,
+        turn_number: int,
+    ) -> None:
+        """Invoke the checkpoint callback if configured.
+
+        Errors are logged but never propagated -- checkpointing must
+        not interrupt execution.
+        """
+        if self._checkpoint_callback is None:
+            return
+        try:
+            await self._checkpoint_callback(ctx)
+        except MemoryError, RecursionError:
+            raise
+        except Exception as exc:
+            logger.exception(
+                EXECUTION_CHECKPOINT_CALLBACK_FAILED,
+                execution_id=ctx.execution_id,
+                turn=turn_number,
+                error=f"{type(exc).__name__}: {exc}",
+            )
+
+    @staticmethod
+    def _finalize(
+        result: ExecutionResult,
+        all_plans: list[ExecutionPlan],
+        replans_used: int,
+    ) -> ExecutionResult:
+        """Attach plan metadata to the execution result."""
+        metadata = copy.deepcopy(result.metadata)
+        metadata.update(
+            {
+                "loop_type": "plan_execute",
+                "plans": [p.model_dump() for p in all_plans],
+                "final_plan": (all_plans[-1].model_dump() if all_plans else None),
+                "replans_used": replans_used,
+            }
+        )
+        return result.model_copy(update={"metadata": metadata})
